+++ src/components/Button/index.ts
@@ -1,4 +1,4 @@
-// Button from copy-ui @ 2025-11-26.
+// Button from copy-ui @ 2025-12-13.
 
 export type { ButtonProps } from "./button";
 export { Button } from "./button";